:recycle: refactor : 전체적인 코드 수정
필요 없는 Dto 삭제, Interface -> implement 구조 삭제, 중복되는 코드 함수화
