+++ src/main/java/com/c_comachi/inused/domain/users/service/UserService.java
@@ -2,15 +2,41 @@ package com.c_comachi.inused.domain.users.service;
 
 import com.c_comachi.inused.domain.users.dto.UserEditInfo;
 import com.c_comachi.inused.domain.users.dto.response.GetLoginUserResponseDto;
+import com.c_comachi.inused.domain.users.entity.UserEntity;
+import com.c_comachi.inused.domain.users.repository.UserRepository;
+import com.c_comachi.inused.global.common.ErrorCode;
+import com.c_comachi.inused.global.exception.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
-public interface UserService {
+@Service
+@RequiredArgsConstructor
+public class UserService{
 
-    ResponseEntity<? super GetLoginUserResponseDto> getLoginUser(String email);
+    private  final UserRepository userRepository;
 
-    void editUser(String email, UserEditInfo requestDto);
+    public ResponseEntity<? super GetLoginUserResponseDto> getLoginUser(String email) {
 
-    void deleteUser(String email);
+        UserEntity userEntity = null;
 
+        userEntity = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
+
+        return GetLoginUserResponseDto.success(userEntity);
+    }
+
+    public void editUser(String email, UserEditInfo info) {
+        UserEntity user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
+        user.userEdit(info);
+        userRepository.save(user);
+    }
+
+    public void deleteUser(String email) {
+        UserEntity userEntity = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
+        userRepository.delete(userEntity);
+    }
 
 }
